requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.6.1
asyncprawcore==2.3.0
attrs==22.1.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
distlib==0.3.6
dnspython==2.2.1
filelock==3.8.2
frozenlist==1.3.3
identify==2.5.9
idna==3.4
motor==3.1.1
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
platformdirs==2.6.0
pre-commit==2.20.0
pymongo==4.3.3
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.13
virtualenv==20.17.1
yarl==1.8.2
